test(cosmovisor): improve TestParseUpgradeInfoFile with substring assertions
In trying to investigate the flake reported in #18073, this change
adds the context of the failing test's filepath. This change allows
us to tighten the tests by using assertions on the error instead of
just checking that it is non-nil.
Also one of the tests that claimed it was empty actually had a
space/newline which meant that it wasn't testing the right expectations.

Updates #18073